pyproject.toml
[build-system]
requires = ["setuptools>=65"]
build-backend = "setuptools.build_meta"

[project]
name = "polymarket-news-trader"
version = "0.1.0"
description = "Event-driven Polymarket trading bot powered by news feeds"
requires-python = ">=3.9"
dependencies = [
    "feedparser>=6.0.0",
    "requests>=2.28.0",
    "PyYAML>=6.0",
]

[project.scripts]
polymarket-news-trader = "polymarket_news_trader.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
